Fix test test_nothing_copies
The original test is not quite right, because when the embargo hours
falls right on an exposure, it should unembargo it.

Copy the original test to a new test test_copy_just_one.

shift the "now" time test_nothing_copies to be 0.000001 second earlier,
so all data still stay within the embargo period, and nothing should
copy.

5596964.255774 sec is the time difference between 2020-01-17 16:55:11.322700
and 2019-11-13T22:12:27.066926, the start time of exposure 2019111300059.